Update ru.ts
I changed the name of some buildings and resources because they didn't fit in the panel.
Also changed the name of the powerbank, as it was called the same as battery, which confused people.
Also in the line 955 does not write the time after which you can change the name. It writes %{час} . I hope I corrected it correctly. If not. Then do not take it.